requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas
numpy
scikit-learn==1.3.2
ta-lib==0.4.25
yfinance==0.2.23
alpaca-trade-api==3.0.2
redis
# sqlite3  # Built into Python
requests==2.31.0
aiohttp
python-telegram-bot==20.7
schedule==1.2.0
plotly==5.17.0
python-dotenv==1.0.0